requirements.txt
Flask==3.1.0
SQLAlchemy==2.0.36
python-dotenv==1.0.1
requests==2.32.3
APScheduler==3.10.4
APScheduler==3.10.4
